### 2020-12-04 #2883 @whedon
[PRE REVIEW]: rgugik: Search and Retrieve Spatial Data from the Polish Head Office of Geodesy and Cartography in R
**Submitting author:** @kadyb (<a href="http://orcid.org/0000-0002-8614-3816">Krzysztof Dyba</a>)
**Repository:** <a href="https://github.com/kadyb/rgugik" target ="_blank">https://github.com/kadyb/rgugik</a>
**Version:** 0.2.0
