### 2020-12-04 #2883 @kyleniemeyer
Hi @kadyb, thanks for your interest in JOSS. The editorial board is going to review your submission to see if it falls without our scope—in particular, this looks like a utility/API package. (See our definition of [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) for more info.)
