+++ pkgdown.yml
@@ -6,5 +6,5 @@ articles:
   articles/spatialdb_description: spatialdb_description.html
   orthophotomap: orthophotomap.html
   topodb: topodb.html
-last_built: 2020-12-03T19:00Z
+last_built: 2020-12-05T11:35Z
 
