### 2020-12-05 #2883 @kadyb
Hello @kyleniemeyer, thank you for your reply. Please let us better explain why we fit the scope of the journal. We believe that our software meets the research software requirements, and it is no small utility or a simple API client. rgugik integrates multiple data sources of different formats and allows them to be used natively in R, making the researcher's work much easier and faster. The package allows searching for, automatically retrieve, and preprocess data in one place, which is currently the most significant obstacle. Most importantly, it supports the complete reproducible research regardless of the goal of the study. Due to the diversity of available datasets, this package can be vital in addressing social, economic, or environmental research challenges and be an integral part of interdisciplinary studies. A larger outline of the needs can be found in the submitted paper. We want to add that we have a certain user base, where the majority are scientists from various fields, including forestry, ecology, and economics.
