### 2020-12-07 9b61c02
fix fun name in DESCRIPTION
