+++ DESCRIPTION
@@ -27,7 +27,7 @@ Description: Automatic open data acquisition from resources of Polish Head Offic
     such as orthophotomaps, Digital Elevation Models (DTM, DSM, Point Cloud),
     State Register of Borders, spatial databases, geometries of cadastral parcels,
     3D models of buildings in LOD1 and LOD2 standards, and more.
-    It is also possible to geocode addresses or objects using the "geocodePL_get"
+    It is also possible to geocode addresses or objects using the geocodePL_get()
     function.
 License: MIT + file LICENSE
 Depends: 
