### 2020-12-07 a9b883d
fix text in DESCRIPTION
