+++ pkgdown.yml
@@ -6,5 +6,5 @@ articles:
   articles/spatialdb_description: spatialdb_description.html
   orthophotomap: orthophotomap.html
   topodb: topodb.html
-last_built: 2020-12-07T10:25Z
+last_built: 2020-12-07T10:36Z
 
