### 2020-12-07 57b3378
set old wd in 'PRECOMPILE.R'
