### 2020-12-09 3b34e67
cleans and updates readme
