+++ README.md
@@ -17,26 +17,26 @@ MIT](https://img.shields.io/badge/License-MIT-blue.svg)](https://opensource.org/
 [Polish Head Office of Geodesy and Cartography](http://www.gugik.gov.pl)
 including:
 
-  - [Orthophotomaps](http://www.gugik.gov.pl/pzgik/zamow-dane/ortofotomapa)
-  - [General Geographic
+-   [Orthophotomaps](http://www.gugik.gov.pl/pzgik/zamow-dane/ortofotomapa)
+-   [General Geographic
     Database](http://www.gugik.gov.pl/pzgik/zamow-dane/baza-danych-obiektow-ogolnogeograficznych)
-  - [Topographic
+-   [Topographic
     Database](http://www.gugik.gov.pl/pzgik/zamow-dane/baza-danych-obiektow-topograficznych-bdot-10k)
-  - [Register of Towns, Streets and
+-   [Register of Towns, Streets and
     Addresses](https://emuia.gugik.gov.pl)
-  - [State Register of Geographical
+-   [State Register of Geographical
     Names](http://www.gugik.gov.pl/pzgik/zamow-dane/panstwowy-rejestr-nazw-geograficznych)
-  - [State Register of
+-   [State Register of
     Borders](http://www.gugik.gov.pl/pzgik/zamow-dane/panstwowy-rejestr-granic-i-powierzchni-jednostek-podzialow-terytorialnych-kraju)
-  - Location (geometry) of cadastral parcels using TERYT (parcel ID) or
+-   Location (geometry) of cadastral parcels using TERYT (parcel ID) or
     coordinates
-  - 3D models of buildings (LOD1, LOD2)
-  - Various Digital Elevation Models as:
-      - [Digital Terrain
+-   3D models of buildings (LOD1, LOD2)
+-   Various Digital Elevation Models as:
+    -   [Digital Terrain
         Model](http://www.gugik.gov.pl/pzgik/zamow-dane/numeryczny-model-terenu)
-      - [Digital Surface
+    -   [Digital Surface
         Model](http://www.gugik.gov.pl/pzgik/zamow-dane/numeryczny-model-pokrycia-terenu)
-      - [Point
+    -   [Point
         Cloud](http://www.gugik.gov.pl/pzgik/zamow-dane/dane-pomiarowe)
 
 It is also possible to geocode addresses or objects using the
@@ -45,7 +45,7 @@ It is also possible to geocode addresses or objects using the
 **Corresponding functions**
 
 | Function                              | Input                  | Dastaset EN                              | Dataset PL                                |
-| :------------------------------------ | :--------------------- | :--------------------------------------- | :---------------------------------------- |
+|:--------------------------------------|:-----------------------|:-----------------------------------------|:------------------------------------------|
 | `orto_request()`, `tile_download()`   | polygon                | Orthophotomap                            | Ortofotomapa                              |
 | `geodb_download()`                    | voivodeship            | General Geographic Database              | Baza Danych Obiektów Ogólnogeograficznych |
 | `topodb_download()`                   | county                 | Topographic Database                     | Baza Danych Obiektów Topograficznych      |
@@ -58,18 +58,18 @@ It is also possible to geocode addresses or objects using the
 
 There are the additional functions for obtaining Digital Terrain Model:
 
-  - `pointDTM_get()` for small areas (high resolution grid)
-  - `pointDTM100_download()` for voivodeships areas (low resolution
+-   `pointDTM_get()` for small areas (high resolution grid)
+-   `pointDTM100_download()` for voivodeships areas (low resolution
[…]
