### 2020-12-12 #2883 @Kevin-Mattheus-Moerman
@kadyb the board has reviewed this work for scope and decided this work may be in scope for JOSS. I will therefore proceed to find a handling editor for this submission.
