+++ tests/testthat/test-tile_download.R
@@ -2,11 +2,13 @@ base_URL = "https://opendata.geoportal.gov.pl/"
 
 sample_orto = paste0(base_URL, "ortofotomapa/41/41_3756_N-33-130-D-b-2-3.tif")
 orto = data.frame(URL = sample_orto, filename = "sample_orto",
-                  sha1 = "312c81963a31e268fc20c442733c48e1aa33838f")
+                  sha1 = "312c81963a31e268fc20c442733c48e1aa33838f",
+                  stringsAsFactors = FALSE)
 
 sample_DEM = paste0(base_URL, "NumDaneWys/NMT/73556/73556_1002897_N-33-130-D-b-2-3.asc")
 DEM = data.frame(URL = sample_DEM, filename = "sample_DEM",
-                 sha1 = "392a0edf763e38fa1b7b6067ac1b080c47374dd2")
+                 sha1 = "392a0edf763e38fa1b7b6067ac1b080c47374dd2",
+                 stringsAsFactors = FALSE)
 
 # ORTO
 tmp = tempfile()
