### 2020-12-14 b940be3
skip equal test in 'minmaxDTM_get'
