### 2020-12-14 #2883 @kadyb
@kbarnhart, thanks for the good news!

We can suggest reviewers from:
a) JOSS list - tscheypidi, jhollist, nuest, mikerspencer,
b) rOpenSci - adamhsparks, boshek
