### 2021-01-04 #2883 @kbarnhart
@kadyb a quick update to to let you know that I've returned to the office and will now start actively handling this submission. I will start to solicit reviewers shortly. Two minor points to address are in the following checklist.

- [ ] It looks like the package has two license files, "LICENSE" and "LICENSE.md". Please update the repo so there is just one plain text license file. Additional information about licenses can be found [on the JOSS docs](https://joss.readthedocs.io/en/latest/review_criteria.html#software-license).
- [ ] On line 54 of the manuscript the reference to `remotes` should not be in the interior set of parentheses. See [this guide](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#Citation_Syntax) regarding rmarkdown citation syntax.

If you have any questions or clarifications, please let me know.
