### 2021-01-05 #2883 @kadyb
Hello, thanks for the answers and suggestions!

> 1. It looks like the package has two license files, "LICENSE" and "LICENSE.md". Please update the repo so there is just one plain text license file. Additional information about licenses can be found on the JOSS docs.

We have two license files due to the [CRAN policy](https://www.r-project.org/Licenses/MIT). Examples of other packages are [dplyr](https://github.com/tidyverse/dplyr) or [openssl](https://github.com/jeroen/openssl).

> 2. On line 54 of the manuscript the reference to remotes should not be in the interior set of parentheses. See this guide regarding rmarkdown citation syntax.

Fixed.

> 3. As part of the review process, reviewers will need to assess whether it is possible to test the software to verify it has been installed/compiled correctly. I would recommend adding instructions describing how a user might do do this.

The package can be installed in the standard way from CRAN or the development version from GitHub. Instructions are available in [README]( https://github.com/kadyb/rgugik#installation). Moreover, examples of using the package are available in [docs]( https://kadyb.github.io/rgugik/reference/index.html) and [vignettes]( https://kadyb.github.io/rgugik/articles/).
