### 2021-01-07 #2883 @kbarnhart
@kadyb thanks for those updates and clarification regarding the license files.

@adamhsparks thank you for your willingness to review.
