### 2021-02-19 kadyb/rgugik#55 @adamhsparks
R documentation format
Related to my review, https://github.com/openjournals/joss-reviews/issues/2948,

The Guidelines for Rd files state that `\title` sections should be capitalised and not end in a period, see: <https://developer.r-project.org/Rds.html>

Currently, the ROxygen sections of the .R files are in this format:
`first Second Third Word` or `first second third word`

It should instead be:
`First second third word` unless there are proper names in the title for each function being documented.
