### 2021-02-19 kadyb/rgugik#55 @adamhsparks
Further comments on the documentation format.

The https://developer.r-project.org/Rds.html guidelines states:

>Always use <- rather than = for assignments.

This example shows an incorrect usage of `=` in place of `<-`.

```r
#' @examples
#' \dontrun{
#' voivodeship_geom = borders_get(voivodeship = "lubuskie") # 494 KB
#' county_geom = borders_get(county = "Sopot") # 18 KB
#' commune_geom = borders_get(commune = c("Hel", "Krynica Morska")) # 11 KB
#' }
```
