### 2021-02-19 kadyb/rgugik#55 @adamhsparks
The term "Boolean" is misspelled in the documentation as "boolean" in at least one spot that I caught.
