### 2021-02-19 #2948 @adamhsparks
Overall the package is well written and I'm impressed with the real-world use-case examples in the vignettes. This will be a great benefit to the community by providing an easy to use resource for a reproducible workflow that centralises data fetching for this data resource. Other tools that offer access to this resource are discussed, along with the advantages of this tool over them.

The contribution guidelines are some of the most clear I've seen. Well done!

I've opened a few issues in the repository for the code or documentation issues I've found or other thoughts.

Here are some more specific comments on the paper and package.
1. I think that the paper would benefit from some edits for clarity and flow, so I've not ticked that off yet. I found the writing to be mostly clear, all of the content was covered as required, but in some places the language was awkward and hard to follow.
2. Two functions have zero test coverage, `geonames_download()` and `borders_download()`, the reason for this isn't clear to me.
3. I noted that the test coverage in the README badge, ~87% and paper, don't match what I get locally, ~80%. I'm unsure of the reasons for such a discrepancy?
4. There are four vignettes in the pkgdown website, but only three in the package as installed from the master branch?
5. In several places "digital elevation model" is spelled as "Digital Elevation Model", this is not correct, it is not a proper noun.
6. The package passes all the normal R CMD checks and whatnot and has even used CI to run lintr on the code, which really shows effort put into writing good code. 🙏
