### 2021-02-20 767b9f1
use 'basename()' instead 'idx_name' var
