+++ NAMESPACE
@@ -9,6 +9,7 @@ export(geodb_download)
 export(geonames_download)
 export(minmaxDTM_get)
 export(models3D_download)
+export(ortho_request)
 export(orto_request)
 export(parcel_get)
 export(pointDTM100_download)
