+++ R/tile_download.R
@@ -24,7 +24,7 @@
 #' polygon_path = system.file("datasets/search_area.gpkg", package = "rgugik")
 #' polygon = read_sf(polygon_path)
 #'
-#' req_df = orto_request(polygon)
+#' req_df = ortho_request(polygon)
 #' tile_download(req_df[1, ]) # download the first image only
 #'
 #' req_df = DEM_request(polygon)
