+++ README.Rmd
@@ -62,7 +62,7 @@ ds_en = c("Orthophotomap",
           "3D models of buildings",
           "Digital Elevation Models")
 
-fun = c("`orto_request()`, `tile_download()`",
+fun = c("`ortho_request()`, `tile_download()`",
         "`geodb_download()`",
         "`topodb_download()`",
         "`emuia_download()`",
@@ -117,10 +117,10 @@ remotes::install_github("kadyb/rgugik")
 
 ## Usage
 
-### Ortophotomap
+### Orthophotomap
   
-- `orto_request()` - returns a data frame with metadata and links to the orthoimages in a given polygon
-- `tile_download()` - downloads orthoimages based on the data frame obtained using the `orto_request()` function
+- `ortho_request()` - returns a data frame with metadata and links to the orthoimages in a given polygon
+- `tile_download()` - downloads orthoimages based on the data frame obtained using the `ortho_request()` function
   
 ```{r f1, message=FALSE, warning=FALSE}
 library(rgugik)
@@ -130,7 +130,7 @@ library(raster)
 polygon_path = system.file("datasets/search_area.gpkg", package = "rgugik")
 polygon = read_sf(polygon_path)
 
-req_df = orto_request(polygon)
+req_df = ortho_request(polygon)
 
 # print metadata of the first image
 t(req_df[1, ])
