+++ vignettes/DEM.Rmd
@@ -113,7 +113,7 @@ The exact area can be determined from the polygon layer as in [orthophotomap](ht
 
 ### Raster data
 
-Now we can search for available elevation data for this area using the `DEM_request()` function (it is analogous to the `orto_request()` function).
+Now we can search for available elevation data for this area using the `DEM_request()` function (it is analogous to the `ortho_request()` function).
 The only argument of the function is our reserve buffer.
 Note that `DEM_request()` takes data frame as input, and currently the `morasko_buffer` object is a list, so we need to convert it first.
 
@@ -132,17 +132,28 @@ req_df[1:10, 1:5]
 ```
 
 ```
-##                 sheetID year              format resolution avgElevErr
-## 1        N-33-130-D-b-1 2011           ASCII TBD     25.0 m        0.9
-## 2      N-33-130-D-b-1-1 2012      ASCII XYZ GRID      0.5 m        0.1
-## 3      N-33-130-D-b-1-1 2012 ARC/INFO ASCII GRID      0.5 m        0.1
-## 4      N-33-130-D-b-1-1 2012      ASCII XYZ GRID      1.0 m        0.1
-## 5      N-33-130-D-b-1-1 2012 ARC/INFO ASCII GRID      1.0 m        0.1
-## 6  N-33-130-D-b-1-1-1-4 2012                 LAS    12 p/m2        0.1
-## 7  N-33-130-D-b-1-1-2-3 2012                 LAS    12 p/m2        0.1
-## 8  N-33-130-D-b-1-1-2-4 2012                 LAS    12 p/m2        0.1
-## 9  N-33-130-D-b-1-1-3-2 2012                 LAS    12 p/m2        0.1
-## 10 N-33-130-D-b-1-1-4-1 2012                 LAS    12 p/m2        0.1
+##                 sheetID year              format resolution
+## 1        N-33-130-D-b-1 2011           ASCII TBD     25.0 m
+## 2      N-33-130-D-b-1-1 2012      ASCII XYZ GRID      0.5 m
+## 3      N-33-130-D-b-1-1 2012 ARC/INFO ASCII GRID      0.5 m
+## 4      N-33-130-D-b-1-1 2012      ASCII XYZ GRID      1.0 m
+## 5      N-33-130-D-b-1-1 2012 ARC/INFO ASCII GRID      1.0 m
+## 6  N-33-130-D-b-1-1-1-4 2012                 LAS    12 p/m2
+## 7  N-33-130-D-b-1-1-2-3 2012                 LAS    12 p/m2
+## 8  N-33-130-D-b-1-1-2-4 2012                 LAS    12 p/m2
+## 9  N-33-130-D-b-1-1-3-2 2012                 LAS    12 p/m2
+## 10 N-33-130-D-b-1-1-4-1 2012                 LAS    12 p/m2
+##    avgElevErr
+## 1         0.9
+## 2         0.1
+## 3         0.1
+## 4         0.1
+## 5         0.1
+## 6         0.1
+## 7         0.1
+## 8         0.1
+## 9         0.1
+## 10        0.1
 ```
 
 We have received metadata with many types of data of different formats, timeliness, resolution, and accuracy.
@@ -241,7 +252,7 @@ st_dimensions(DEM)
 ```
 
 ```
-##      from   to offset delta               refsys point   values    
+##      from   to offset delta               refsys point   values x/y
 ## x       1 2188 355733     1 ETRS89 / Poland CS92  TRUE     NULL [x]
 ## y       1 2379 517029    -1 ETRS89 / Poland CS92  TRUE     NULL [y]
 ## elev    1    2     NA    NA                   NA    NA DTM, DSM
