+++ R/tile_download.R
@@ -1,7 +1,7 @@
 #' downloads tiles based on the data frame obtained using
-#' the orto_request() and DEM_request() functions
+#' the ortho_request() and DEM_request() functions
 #'
-#' @param df_req a data frame obtained using the [`orto_request()`] and
+#' @param df_req a data frame obtained using the [`ortho_request()`] and
 #' [`DEM_request()`] functions
 #' @param outdir (optional) name of the output directory;
 #' by default, files are saved in the working directory
@@ -34,11 +34,11 @@ tile_download = function(df_req, outdir = ".", unzip = TRUE, check_SHA = FALSE,
                          print_iter = TRUE, ...) {
 
   if (!"URL" %in% names(df_req)) {
-    stop("data frame should come from 'request_orto'")
+    stop("data frame should come from 'request_ortho'")
   }
 
   if (!"filename" %in% names(df_req)) {
-    stop("data frame should come from 'request_orto'")
+    stop("data frame should come from 'request_ortho'")
   }
 
   if (check_SHA == TRUE && !"sha1" %in% names(df_req)) {
