### 2021-02-20 1eba41d
rename boolean to logical
