+++ R/county_names.R
@@ -2,8 +2,8 @@
 #' @title Counties in Poland
 #'
 #' @description The data frame contains the names of counties,
-#' their identifiers (TERYT, 4 characters) and the availability of 
-#' building models in the LOD2 standard (boolean value).
+#' their identifiers (TERYT, 4 characters) and the availability of
+#' building models in the LOD2 standard (logical value).
 #'
 #' @importFrom utils data
 #'
