### 2021-02-21 kadyb/rgugik#55 @kadyb
> The Guidelines for Rd files state that \title sections should be capitalised and not end in a period, see: https://developer.r-project.org/Rds.html

Fixed in https://github.com/kadyb/rgugik/pull/59.
