### 2021-02-21 kadyb/rgugik#55 @kadyb
> Always use <- rather than = for assignments.

If possible, we would prefer to stay with the `=` for the assignment to be consistent with our code and vignettes. In my opinion, more and more users and developers are using `=` instead of the `<-`, which makes writing code much easier. This has a great advantage when other languages, such as C++ or Python, are used in the project in addition to R.
