### 2021-02-22 #2948 @kbarnhart
@adamhsparks thanks for your review. @kadyb if discussion is necessary regarding any of the comments I'd recommend making an issue in the main repository. Otherwise you can respond here.

@mikerspencer a friendly reminder to provide an update on your review.

Thanks again to both of you for contributing to the JOSS review process.
