### 2021-02-22 #2948 @kbarnhart
@mikerspencer thanks for letting me know. At JOSS, we aim to be very understanding that the pandemic has changed people's time availability (e.g. , [see this blog post from early in the pandemic](https://blog.joss.theoj.org/2020/05/reopening-joss)). Just keep me posted if you need more time beyond this week.
