### 2021-02-23 kadyb/rgugik#55 @adamhsparks
To be clear, this issue is not about using it in the package but rather in the examples in the documentation itself where it's unconventional to use `=`. But as I said, it's not hard-fast. Just more normal for R, which as you noted is not Python or C++.
