### 2021-02-23 kadyb/rgugik#61 @mikerspencer
2x licence files
There's perhaps a good reason, but two licence files could easily be confusing. e.g. Github identifies the `LICENCE` file, which contains the authors and a copyright statement.

Perhaps a CITATION file would help?

As part of JOSS review: https://github.com/openjournals/joss-reviews/issues/2948
