### 2021-02-23 kadyb/rgugik#62 @mikerspencer
JOSS paper authors
There are two JOSS paper authors, but the package description file contains 4 authors.

As part of JOSS review: https://github.com/openjournals/joss-reviews/issues/2948
