### 2021-02-23 kadyb/rgugik#63 @mikerspencer
Link to cran page/vignettes
Consider adding a link in the `readme` to the project cran page, or adding links to the vignettes. Many folk are introduced to a package through a vignette, so raising their profile/visibility will help increase your package uptake.

Submitted as part of JOSS review: https://github.com/openjournals/joss-reviews/issues/2948.
