### 2021-02-23 kadyb/rgugik#64 @mikerspencer
Connection error in tile download
When running `tile_download(req_df, outdir = "./data")` from vignette https://cran.r-project.org/web/packages/rgugik/vignettes/DEM.html I get a connection error after ~15 MB.

I'll try again in a few hours, in case the problem is my end.
