### 2021-02-23 kadyb/rgugik#64 @mikerspencer
Same error is occurring. Of the 33.7 MB file I get 15, 16.7 & 17.8 MB. I think it's too consistent for it to be a problem with my internet connection.
