### 2021-02-23 kadyb/rgugik#64 @mikerspencer
Same connection error when downloading other files, e.g. running `geodb_download("wielkopolskie", outdir = "./data")` from https://cran.r-project.org/web/packages/rgugik/vignettes/orthophotomap.html
