### 2021-02-23 kadyb/rgugik#64 @mikerspencer
Error also on `topodb_download("bieszczadzki", outdir = "./data")`. Again at ~15 MB downloaded.
