### 2021-02-23 kadyb/rgugik#64 @mikerspencer
The first 2.4 MB file from `test-topodb_download.R` downloads, but larger files fail at ~15 MB.
