### 2021-02-23 kadyb/rgugik#64 @mikerspencer
I suggest adding a larger file to the download tests to catch the `connection error`.
