### 2021-02-23 #2948 @mikerspencer
@kbarnhart I've finished my review. There's a lot to commend, especially how clearly the package is structured and documented.

I've raised a couple of minor issues (#61, #62 & #63), but unfortunately I was unable to test functionality due to issue #64. Issue #64 is a connection error when downloading datasets, which consistently occurs at ~15 MB through a download. The issue doesn't affect smaller downloads, e.g. the 2.4 MB files in the automated tests. I've tested at two different times of day and am reasonably sure my internet connection is stable. I'm unsure what to recommend, but have suggested adding larger files to the automated download tests.
