### 2021-02-24 kadyb/rgugik#64 @kadyb
Sorry for that. I will investigate it. Do you use LAN or WiFi connection?

Hello @adamhsparks, as you are a foreign user, did you also encounter this problem during the review?
