### 2021-02-24 kadyb/rgugik#64 @adamhsparks
Nothing I detected, but I’ll try. My Australian Internet connection should be a true test. 😂
