### 2021-02-24 kadyb/rgugik#64 @adamhsparks
`geodb_download("wielkopolskie", outdir = "~/tmp")` and `topodb_download("bieszczadzki", outdir = "~/tmp")` both work fine for me over WiFi with my crappy connection in Perth.
