### 2021-02-24 kadyb/rgugik#64 @kadyb
Thanks for the test! Could you additionally check one large file (375MB) to be sure?

```
borders_download("administrative units")
```
