### 2021-02-24 kadyb/rgugik#64 @adamhsparks
ah, here we go.

```r
> # combine data tables
> req_df = rbind(req_df_DTM, req_df_DSM)
> req_df[, 1:5]
            sheetID year              format resolution avgElevErr
30 N-33-130-D-b-1-1 2019 ARC/INFO ASCII GRID      1.0 m        0.1
29 N-33-130-D-b-1-1 2019 ARC/INFO ASCII GRID      0.5 m        0.1
> tile_download(req_df, outdir = "./data")
1/2
trying URL 'https://opendata.geoportal.gov.pl/NumDaneWys/NMT/73044/73044_917579_N-33-130-D-b-1-1.asc'
Content type 'application/octet-stream' length 35303843 bytes (33.7 MB)
==================================================
downloaded 33.7 MB

2/2
trying URL 'https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc'
Content type 'application/octet-stream' length 141586478 bytes (135.0 MB)
===================
downloaded 53.5 MB

[1] "connection error"
```
