### 2021-02-24 kadyb/rgugik#64 @adamhsparks
``` r
borders_download("administrative units")
#> Error in borders_download("administrative units"): could not find function "borders_download"
```
