### 2021-02-24 kadyb/rgugik#64 @adamhsparks
```r
> borders_download("administrative units")
trying URL 'ftp://91.223.135.109/prg/jednostki_administracyjne.zip'
Content type 'unknown' length 393405215 bytes (375.2 MB)
==
[1] "connection error"
```
