### 2021-02-24 kadyb/rgugik#64 @adamhsparks
I do frequently have to adjust the timeout in my R session when using `download.file()`. I'm checking right now using wget with the administrative units. It says it will take ~10 minutes or so. Guessing we're seeing a timeout here in `download.files()`.
