### 2021-02-24 kadyb/rgugik#64 @mikerspencer
> Sorry for that. I will investigate it. Do you use LAN or WiFi connection?
>

I'm on a wired, 1 GB/s, connection to my router. It's not a particularly fast internet connection (~18 MB/s), but it is stable.
