### 2021-02-24 kadyb/rgugik#64 @adamhsparks
No.

```r
> options(timeout=10000)
> utils::download.file(URL, filename, mode = "wb")
trying URL 'ftp://91.223.135.109/prg/jednostki_administracyjne.zip'
Content type 'unknown' length 393405215 bytes (375.2 MB)
==================================================
Error in utils::download.file(URL, filename, mode = "wb") :
  cannot open URL 'ftp://91.223.135.109/prg/jednostki_administracyjne.zip'
In addition: Warning message:
In utils::download.file(URL, filename, mode = "wb") :
  URL 'ftp://91.223.135.109/prg/jednostki_administracyjne.zip': status was 'Failure when receiving data from the peer'
```
