### 2021-02-24 kadyb/rgugik#64 @adamhsparks
BTW, wget worked fine using zsh
