### 2021-02-24 kadyb/rgugik#64 @kadyb
Thanks for the responses. Here are the logs how it looks for me:

```r
> system.time(borders_download("administrative units"))
trying URL 'ftp://91.223.135.109/prg/jednostki_administracyjne.zip'
downloaded 375.2 MB

   user  system elapsed
  23.34   90.65  949.96

> system.time(geodb_download("wielkopolskie"))
trying URL 'http://opendata.geoportal.gov.pl/bdoo/PL.PZGiK.201.30.zip'
Content type 'application/octet-stream' length 18436314 bytes (17.6 MB)
downloaded 17.6 MB

   user  system elapsed
   1.76    3.04   22.17

> system.time(topodb_download("bieszczadzki"))
trying URL 'https://opendata.geoportal.gov.pl/bdot10k/18/1801_GML.zip'
Content type 'application/octet-stream' length 23474972 bytes (22.4 MB)
downloaded 22.4 MB

   user  system elapsed
   2.34    4.69   24.31

> system.time(tile_download(req_df))
1/2
trying URL 'https://opendata.geoportal.gov.pl/NumDaneWys/NMT/73044/73044_917579_N-33-130-D-b-1-1.asc'
Content type 'application/octet-stream' length 35303843 bytes (33.7 MB)
downloaded 33.7 MB

2/2
trying URL 'https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc'
Content type 'application/octet-stream' length 141586478 bytes (135.0 MB)
downloaded 135.0 MB

   user  system elapsed
   4.03   25.47  134.23
```
