### 2021-02-24 kadyb/rgugik#64 @kadyb
There is clearly a problem. Perhaps users from outside Poland have lower connection priority and therefore disconnects occur?
All functions in the package are wrappers to R `utils::download.file()`.
I'll check a few more things.
