### 2021-02-24 kadyb/rgugik#64 @adamhsparks
Not sure but it seems an issue in R. Like I said, using wget worked fine for me in the terminal (not RStudio, iTerm2), it downloaded a file that R wouldn’t.

`curl::curl_download()` may be useful here providing different handle options that `utils::download.file()` can’t/doesn’t provide?
