### 2021-02-24 kadyb/rgugik#64 @kadyb
I have some results. The problem appears to be exceptional and difficult to reproduce. The following code works for me without any problems. I use Windows 8.1.

```r
test1 = 'https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc'
test2 = 'https://integracja.gugik.gov.pl/Budynki3D/pobierz.php?plik=powiaty/lod1/3064_gml.zip'

tmp = tempfile()
for (i in seq_len(30)) {
  utils::download.file(test1, tmp, method = "wininet", mode = "wb")
  utils::download.file(test1, tmp, method = "libcurl", mode = "wb")

  utils::download.file(test2, tmp, method = "wininet", mode = "wb")
  utils::download.file(test2, tmp, method = "libcurl", mode = "wb")
}
```

By moving away from WiFi I can trigger a timeout error:

```r
>   utils::download.file(test1, tmp, method = "libcurl", mode = "wb")
trying URL 'https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc'
 length 141586478 bytes (135.0 MB)
downloaded 118.1 MB

Error in utils::download.file(test1, tmp, method = "libcurl", mode = "wb") :
  download from 'https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc' failed
In addition: Warning messages:
1: In utils::download.file(test1, tmp, method = "libcurl", mode = "wb") :
  downloaded length 123876510 != reported length 141586478
2: In utils::download.file(test1, tmp, method = "libcurl", mode = "wb") :
  URL 'https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc': Timeout of 60 seconds was reached
```

To fix this, just set `options(timeout = 10000)`. Also, @adamhsparks tried to do this without success, so the actual error is different. From his logs I can see:

> 'Failure when receiving data from the peer'
