### 2021-02-24 kadyb/rgugik#64 @kadyb
Then I tried the solution suggested by @adamhsparks and the strange thing is that `curl::curl_download()` completely fails for me, but `curl::curl_fetch_disk()` works fine.

```r
curl::curl_download(test1, "test.zip", quiet = FALSE)
#>  [-1073741824%] Downloaded 0 bytes...
curl::curl_download(test2, "test.zip", quiet = FALSE)
#>  [-805306368%] Downloaded 0 bytes...

curl::curl_fetch_disk(test1, "test.zip")
file.size("test.zip")/1024^2
#> 135.0274
curl::curl_fetch_disk(test2, "test.zip")
file.size("test.zip")/1024^2
#> 20.76025
```

Could you confirm that `curl::curl_download()` or `curl::curl_fetch_disk()` works for downloading `test1` and `test2` files, please?

CC @Nowosad
