### 2021-02-24 kadyb/rgugik#61 @kadyb
This is a widely known problem and unfortunately nothing can be done about it (see [1](https://stackoverflow.com/questions/43550479/how-to-satisfy-both-cran-and-github-license-file-naming-requirements), [2](https://github.com/tidyverse/tidyr/issues/262)). This is due to the [CRAN policy](https://www.r-project.org/Licenses/MIT).

I tried to use the method mentioned on Stack Overflow in the past, but it doesn't currently work. This issue doesn't have a good solution (now we use the MIT badge in the [README](https://github.com/kadyb/rgugik/blob/master/README.md) so that the license is more visible). The license is also mentioned in the package description on [CRAN](https://cran.r-project.org/web/packages/rgugik/index.html).

As a reference, other packages ([dplyr](https://github.com/tidyverse/dplyr), [openssl](https://github.com/jeroen/openssl)) also have this issue.
