### 2021-02-24 kadyb/rgugik#64 @mikerspencer
OK, attempts here:

> test1 = 'https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc'
> test2 = 'https://integracja.gugik.gov.pl/Budynki3D/pobierz.php?plik=powiaty/lod1/3064_gml.zip'
>
> tic = proc.time()
> curl::curl_download(test1, "test.zip", quiet = FALSE)
 [100%] Downloaded 141586478 bytes...
> proc.time() - tic
   user  system elapsed
  4.061   2.896 421.931
> file.size("test.zip")/1024^2
[1] 135.0274
>
> tic = proc.time()
> curl::curl_download(test2, "test.zip", quiet = FALSE)
 [100%] Downloaded 21768701 bytes...
> proc.time() - tic
   user  system elapsed
  1.516   0.963  28.770
> file.size("test.zip")/1024^2
[1] 20.76025
>
> tic = proc.time()
> curl::curl_fetch_disk(test1, "test.zip")
$url
[1] "https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc"

$times
     redirect    namelookup       connect   pretransfer starttransfer         total
     0.000000      0.061644      0.107014      0.239521      0.621672    551.009517

$content
[1] "/home/mike/test.zip"

> proc.time() - tic
   user  system elapsed
  1.182   1.470 551.036
> file.size("test.zip")/1024^2
[1] 135.0274
>
> tic = proc.time()
> curl::curl_fetch_disk(test2, "test.zip")
$url
[1] "https://integracja.gugik.gov.pl/Budynki3D/pobierz.php?plik=powiaty/lod1/3064_gml.zip"

$times
     redirect    namelookup       connect   pretransfer starttransfer         total
     0.000000      0.075205      0.126705      0.254386      0.330570     28.053150

$content
[1] "/home/mike/test.zip"

> proc.time() - tic
   user  system elapsed
  0.595   0.597  28.094
> file.size("test.zip")/1024^2
[1] 20.76025
