### 2021-02-24 kadyb/rgugik#62 @kadyb
Thanks for bringing this up. It needs clarification. We use [R terminology](https://rdrr.io/r/utils/person.html) to define the roles of individuals in creating the package. Three people and one institution are listed in the package [DESCRIPTION](https://github.com/kadyb/rgugik/blob/master/DESCRIPTION):
* Krzysztof Dyba (`Author`, `Creator`)
* Jakub Nowosad (`Author`)
* Maciej Beręsewicz (`Contributor`)
* *Główny Urząd Geodezji i Kartografii* (`Contributor`)

`Author` and `Creator` made significant contributions and influence the development of the package while the `Contributor` only made small contributions. *Główny Urząd Geodezji i Kartografii* has no contributions, it is only listed as a data source for formality.

Me and Jakub Nowosad are the main developers of the package and we were responsible for writing the article.
Contributions can be checked here: https://github.com/kadyb/rgugik/graphs/contributors
