### 2021-02-24 kadyb/rgugik#63 @kadyb
Thanks! This is a really great idea. I also think the vignettes will be very helpful to novice users so I will add them to the README.

BTW: The CRAN website can be accessed by clicking on the CRAN badge in [README](https://github.com/kadyb/rgugik/blob/master/README.md).
