### 2021-02-24 kadyb/rgugik#63 @kadyb
I fixed this in 04bdc32, but there is a problem with installing **textshaping** package from source (https://github.com/kadyb/rgugik/actions/runs/596692037). It should resolve itself in a few days. I will leave this issue open for now.
