### 2021-02-24 kadyb/rgugik#55 @kadyb
Yeah, I just mean that I would like to keep the same assignment character in all places (including examples). This is mainly due to my habit, as I also use Python. We are aware that a lot of R users prefer the `<-` character. As I mentioned before, we have noticed that more and more users are simply using `=` and perhaps this will be the default assignment character in R in the future.

For references, you can check books focused on spatial data analysis ([Geocomputation with R](https://geocompr.robinlovelace.net/), [Spatial Data Science](https://keen-swartz-3146c4.netlify.app/)), examples in [sf](https://github.com/r-spatial/sf) and [stars](https://github.com/r-spatial/stars) packages or [this](https://github.com/Robinlovelace/geocompr/issues/319#issuecomment-427376764) interesting post.

Nevertheless, thanks for this suggestion and understanding!
