### 2021-02-24 kadyb/rgugik#64 @adamhsparks
My attempts match @mikerspencer's. Everything downloads fine in R using either `curl` function.
