### 2021-02-25 kadyb/rgugik#64 @Nowosad
Both **curl** functions work well on my (Fedora) machine:

``` r
test1 = 'https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc'
test2 = 'https://integracja.gugik.gov.pl/Budynki3D/pobierz.php?plik=powiaty/lod1/3064_gml.zip'

curl::curl_download(test1, "test1.zip", quiet = FALSE)
curl::curl_download(test2, "test2.zip", quiet = FALSE)

curl::curl_fetch_disk(test1, "test1.zip")
#> $url
#> [1] "https://opendata.geoportal.gov.pl/NumDaneWys/NMPT/73043/73043_917495_N-33-130-D-b-1-1.asc"
#>
#> $status_code
#> [1] 200
#>
#> $type
#> [1] "application/octet-stream"
#>
#> $headers
#>   [1] 48 54 54 50 2f 31 2e 31 20 32 30 30 20 4f 4b 0d 0a 44 61 74 65 3a 20 54 68
#>  [26] 75 2c 20 32 35 20 46 65 62 20 32 30 32 31 20 30 38 3a 31 36 3a 34 31 20 47
#>  [51] 4d 54 0d 0a 53 65 72 76 65 72 3a 20 41 70 61 63 68 65 2f 32 2e 34 2e 36 20
#>  [76] 28 43 65 6e 74 4f 53 29 20 4f 70 65 6e 53 53 4c 2f 31 2e 30 2e 32 6b 2d 66
#> [101] 69 70 73 20 50 48 50 2f 37 2e 30 2e 33 33 0d 0a 58 2d 50 6f 77 65 72 65 64
#> [126] 2d 42 79 3a 20 50 48 50 2f 37 2e 30 2e 33 33 0d 0a 43 6f 6e 74 65 6e 74 2d
#> [151] 44 65 73 63 72 69 70 74 69 6f 6e 3a 20 46 69 6c 65 20 54 72 61 6e 73 66 65
#> [176] 72 0d 0a 43 6f 6e 74 65 6e 74 2d 44 69 73 70 6f 73 69 74 69 6f 6e 3a 20 61
#> [201] 74 74 61 63 68 6d 65 6e 74 3b 20 66 69 6c 65 6e 61 6d 65 3d 22 37 33 30 34
#> [226] 33 5f 39 31 37 34 39 35 5f 4e 2d 33 33 2d 31 33 30 2d 44 2d 62 2d 31 2d 31
#> [251] 2e 61 73 63 22 0d 0a 45 78 70 69 72 65 73 3a 20 30 0d 0a 43 61 63 68 65 2d
#> [276] 43 6f 6e 74 72 6f 6c 3a 20 6d 75 73 74 2d 72 65 76 61 6c 69 64 61 74 65 0d
#> [301] 0a 50 72 61 67 6d 61 3a 20 70 75 62 6c 69 63 0d 0a 43 6f 6e 74 65 6e 74 2d
#> [326] 4c 65 6e 67 74 68 3a 20 31 34 31 35 38 36 34 37 38 0d 0a 43 6f 6e 74 65 6e
#> [351] 74 2d 54 79 70 65 3a 20 61 70 70 6c 69 63 61 74 69 6f 6e 2f 6f 63 74 65 74
#> [376] 2d 73 74 72 65 61 6d 0d 0a 0d 0a
#>
#> $modified
#> [1] NA
#>
#> $times
#>      redirect    namelookup       connect   pretransfer starttransfer
#>      0.000000      0.000028      0.000029      0.000090      0.287118
#>         total
#>     10.198358
#>
#> $content
#> [1] "/tmp/RtmpdsONDi/reprex499ff6388a34a/test1.zip"
file.size("test1.zip")/1024^2
#> [1] 135.0274
curl::curl_fetch_disk(test2, "test2.zip")
#> $url
#> [1] "https://integracja.gugik.gov.pl/Budynki3D/pobierz.php?plik=powiaty/lod1/3064_gml.zip"
#>
#> $status_code
#> [1] 200
#>
#> $type
#> [1] "application/octet-stream"
#>
#> $headers
#>   [1] 48 54 54 50 2f 31 2e 31 20 32 30 30 20 4f 4b 0d 0a 44 61 74 65 3a 20 54 68
#>  [26] 75 2c 20 32 35 20 46 65 62 20 32 30 32 31 20 30 38 3a 31 36 3a 35 32 20 47
#>  [51] 4d 54 0d 0a 53 65 72 76 65 72 3a 20 41 70 61 63 68 65 2f 32 2e 34 2e 32 35
#>  [76] 20 28 44 65 62 69 61 6e 29 0d 0a 50 72 61 67 6d 61 3a 20 70 75 62 6c 69 63
#> [101] 0d 0a 45 78 70 69 72 65 73 3a 20 30 0d 0a 43 61 63 68 65 2d 43 6f 6e 74 72
#> [126] 6f 6c 3a 20 70 75 62 6c 69 63 0d 0a 43 6f 6e 74 65 6e 74 2d 44 65 73 63 72
#> [151] 69 70 74 69 6f 6e 3a 20 46 69 6c 65 20 54 72 61 6e 73 66 65 72 0d 0a 43 6f
#> [176] 6e 74 65 6e 74 2d 44 69 73 70 6f 73 69 74 69 6f 6e 3a 20 61 74 74 61 63 68
#> [201] 6d 65 6e 74 3b 20 66 69 6c 65 6e 61 6d 65 3d 22 6c 6f 64 31 2f 33 30 36 34
#> [226] 5f 67 6d 6c 2e 7a 69 70 22 0d 0a 43 6f 6e 74 65 6e 74 2d 54 72 61 6e 73 66
#> [251] 65 72 2d 45 6e 63 6f 64 69 6e 67 3a 20 62 69 6e 61 72 79 0d 0a 43 6f 6e 74
#> [276] 65 6e 74 2d 4c 65 6e 67 74 68 3a 20 32 31 37 36 38 37 30 31 0d 0a 43 6f 6e
#> [301] 74 65 6e 74 2d 54 79 70 65 3a 20 61 70 70 6c 69 63 61 74 69 6f 6e 2f 6f 63
#> [326] 74 65 74 2d 73 74 72 65 61 6d 0d 0a 0d 0a
#>
#> $modified
#> [1] NA
#>
#> $times
#>      redirect    namelookup       connect   pretransfer starttransfer
#>      0.000000      0.001322      0.009389      0.042308      0.072064
#>         total
#>      2.308206
#>
#> $content
#> [1] "/tmp/RtmpdsONDi/reprex499ff6388a34a/test2.zip"
file.size("test2.zip")/1024^2
#> [1] 20.76025
```
