### 2021-02-25 kadyb/rgugik#64 @kadyb
Sorry for the many messages, but this problem is hard to reproduce, and I have one more idea.

All functions in the **rgugik** that download files using `utils::download.file()` support parameter passing by `...`. @adamhsparks mentioned that downloading using wget from shell worked fine. So why not try to use `method = "wget"`? I tested it below, but the most important thing is that it works for you. Other download methods are also available.

If nothing works, I will rewrite it to `curl::curl_download()` because you already confirmed that it works well.

```r
library("rgugik")

# 17 MB
geodb_download("wielkopolskie", method = "wget", quiet = TRUE)

# 22 MB
topodb_download("bieszczadzki", method = "wget", quiet = TRUE)

# 48 MB
models3D_download("Warszawa", method = "wget", quiet = TRUE)

# 375 MB
borders_download("administrative units", method = "wget", quiet = TRUE)
```
