### 2021-02-25 546f492
tests use of curl
