+++ R/pointDTM100_download.R
@@ -6,7 +6,7 @@
 #' by default, files are saved in the working directory
 #' @param unzip TRUE (default) or FALSE, when TRUE the downloaded archive will
 #' be extracted and removed
-#' @param ... additional argument for [`utils::download.file()`]
+#' @param ... additional argument for [`curl::curl_download()`]
 #'
 #' @return text files with X, Y, Z columns (EPSG:2180)
 #'
@@ -61,7 +61,7 @@ pointDTM100_download = function(voivodeships, outdir = ".", unzip = TRUE, ...) {
 
   for (i in seq_len(nrow(df_names))) {
     filename = paste0(outdir, "/", df_names[i, type], ".zip")
-    status = tryGet(utils::download.file(df_names[i, "URL"], filename, mode = "wb", ...))
+    status = tryGet(curl::curl_download(df_names[i, "URL"], filename, mode = "wb", ...))
 
     if (any(status %in% c("error", "warning"))) {
       return("connection error")
