+++ R/geonames_download.R
@@ -6,7 +6,7 @@
 #' by default, files are saved in the working directory
 #' @param unzip TRUE (default) or FALSE, when TRUE the downloaded archive will
 #' be extracted and removed
-#' @param ... additional argument for [`utils::download.file()`]
+#' @param ... additional argument for [`curl::curl_download()`]
 #'
 #' @return a selected data type in the specified format
 #'
@@ -51,7 +51,7 @@ geonames_download = function(type, format = "SHP", outdir = ".", unzip = TRUE, .
   for (i in seq_len(nrow(df))) {
     filename = paste0(outdir, "/", df[i, "type"], "_",
                       df[i, "format"], ".zip")
-    status = tryGet(utils::download.file(df[i, "URL"], filename, mode = "wb", ...))
+    status = tryGet(curl::curl_download(df[i, "URL"], filename, mode = "wb", ...))
 
     if (any(status %in% c("error", "warning"))) {
       return("connection error")
