### 2021-02-25 3d0baf2
fix DEM spelling in README
