+++ .github/workflows/rcmdcheck.yml
@@ -34,6 +34,7 @@ jobs:
 
     - name: Install dependencies
       run: |
+        options("pkgType" = "binary")
         remotes::install_deps(dependencies = TRUE)
         remotes::install_cran("rcmdcheck")
       shell: Rscript {0}
