+++ .github/workflows/lintr.yml
@@ -35,6 +35,7 @@ jobs:
 
     - name: Install dependencies
       run: |
+        options("pkgType" = "binary")
         remotes::install_deps(dependencies = TRUE)
         remotes::install_cran("lintr")
       shell: Rscript {0}
