### 2021-02-28 a42cffe
"Cyfrowe Modele" lowercase in README
