### 2021-03-03 bb89775
fix DEM spelling in 'DEM_request.R'
