### 2021-03-03 ea7fe82
fix DTM spelling in 'pointDTM100_download.R'
