+++ R/pointDTM100_download.R
@@ -1,4 +1,4 @@
-#' @title Download Digital Terrain Models for voivodeships (100 m resolution)
+#' @title Download digital terrain models for voivodeships (100 m resolution)
 #'
 #' @param voivodeships selected voivodeships in Polish or English, or TERC
 #' (function [`voivodeship_names()`] can by helpful)
