### 2021-03-03 fcd54f9
fix DEM spelling in vignettes
