+++ vignettes/DEM.Rmd
@@ -1,9 +1,9 @@
 ---
-title: "Digital Elevation Model"
+title: "Digital elevation model"
 author: "Krzysztof Dyba, Jakub Nowosad"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Digital Elevation Model}
+  %\VignetteIndexEntry{Digital elevation model}
   %\VignetteEngine{knitr::rmarkdown}
   \usepackage[utf8]{inputenc}
 ---
@@ -16,15 +16,15 @@ body {text-align: justify}
 
 ## Definition
 
-**Digital Elevation Models** are models describing the terrain surface.
+**Digital elevation models** are models describing the terrain surface.
 They are created as a result of the processing of aerial photos, laser scanning (LiDAR), geodetic surveying, or radar interferometry (InSAR).
 DEMs are one of the key datasets in Geographic Information Systems (GIS) and constitute the basis for many environmental spatial analyses.
 In addition, they are a source for derived products such as terrain slope and aspect.
 DEM is the general name for a group of models with different characteristics, including:
 
-1. **Digital Terrain Model** (DTM) - numerical representation of the terrain with its morphological forms.
+1. **Digital terrain model** (DTM) - numerical representation of the terrain with its morphological forms.
 This model is devoid of any objects above the ground, such as buildings or trees.
-2. **Digital Surface Model** (DSM) - numerical representation of the terrain, including objects on its surface, such as buildings and trees.
+2. **Digital surface model** (DSM) - numerical representation of the terrain, including objects on its surface, such as buildings and trees.
 
 <center>
 ![DTM-DSM](https://upload.wikimedia.org/wikipedia/commons/thumb/6/6c/DTM_DSM.svg/1024px-DTM_DSM.svg.png){ width=50% }
@@ -132,32 +132,21 @@ req_df[1:10, 1:5]
 ```
 
 ```
-##                 sheetID year              format resolution
-## 1        N-33-130-D-b-1 2011           ASCII TBD     25.0 m
-## 2      N-33-130-D-b-1-1 2012      ASCII XYZ GRID      0.5 m
-## 3      N-33-130-D-b-1-1 2012 ARC/INFO ASCII GRID      0.5 m
-## 4      N-33-130-D-b-1-1 2012      ASCII XYZ GRID      1.0 m
-## 5      N-33-130-D-b-1-1 2012 ARC/INFO ASCII GRID      1.0 m
-## 6  N-33-130-D-b-1-1-1-4 2012                 LAS    12 p/m2
-## 7  N-33-130-D-b-1-1-2-3 2012                 LAS    12 p/m2
-## 8  N-33-130-D-b-1-1-2-4 2012                 LAS    12 p/m2
-## 9  N-33-130-D-b-1-1-3-2 2012                 LAS    12 p/m2
-## 10 N-33-130-D-b-1-1-4-1 2012                 LAS    12 p/m2
-##    avgElevErr
-## 1         0.9
-## 2         0.1
-## 3         0.1
-## 4         0.1
-## 5         0.1
-## 6         0.1
-## 7         0.1
-## 8         0.1
-## 9         0.1
-## 10        0.1
+##                 sheetID year              format resolution avgElevErr
+## 1        N-33-130-D-b-1 2011           ASCII TBD     25.0 m        0.9
+## 2      N-33-130-D-b-1-1 2012      ASCII XYZ GRID      0.5 m        0.1
+## 3      N-33-130-D-b-1-1 2012 ARC/INFO ASCII GRID      0.5 m        0.1
[…]
