+++ paper/paper.md
@@ -24,42 +24,43 @@ bibliography: paper.bib
 # Introduction
 
 Currently, the open data market size is estimated at about 185 billion Euros in the European Union and is expected to grow in the coming years [@huyer2020economic].
-It includes spatial data, which release can cause cost savings and be used to create new, innovative products and services for the benefit of the society, environment, and economy.
+It includes spatial data that can cause cost savings and create new, innovative products and services for the benefit of the society, environment, and economy.
 The public sector is one of the primary providers of vast amounts of valuable spatial data resources.
 
-In Poland, the [Head Office of Geodesy and Cartography](http://www.gugik.gov.pl/) (*Główny Urząd Geodezji i Kartografii*, *GUGiK*) is the central government agency responsible for collecting spatial data.
-Their resources include various datasets, such as orthophotomaps, register of borders, 3D models of buildings, Digital Elevation Models, and point clouds.
-Until July 31, 2020, spatial data acquisition was time-consuming and required filling in forms and paying a fee.
+[The Head Office of Geodesy and Cartography](http://www.gugik.gov.pl/) (*Główny Urząd Geodezji i Kartografii*, *GUGiK*) is the central government agency responsible for collecting spatial data in Poland.
+Their resources include various datasets, such as orthophotomaps, register of borders, 3D models of buildings, digital elevation models, and point clouds.
+Until July 31, 2020, spatial data acquisition was time-consuming, required filling-in forms, and paying a fee.
 However, the recent amendment of the Geodetic and Cartographic Law in Poland in mid-2020 made all of the current and future spatial datasets publicly available.
 
-Poland's spatial data is released on a dedicated website, [Geoportal](https://mapy.geoportal.gov.pl), which allows for its browsing and downloading.
-It is one of the most popular government websites in the country, currently ranked 3rd with 5.49 million unique visits in 2020^[https://widok.gov.pl/statistics/].
+Poland's spatial data is released on a dedicated website, [Geoportal](https://mapy.geoportal.gov.pl), which allows to browse and download it.
+The Geoportal is one of the most popular government websites in the country, currently ranked 3rd with 5.5 million unique visits in 2020^[https://widok.gov.pl/statistics/].
 Although the data is related to Poland's area only, visits from other countries can also be noted (e.g., Germany with 52,000, Great Britain with 40,000, and United States with 15,000 unique visits this year)^[https://widok.gov.pl/statistics/geoportal-krajowy/].
-In the first month after the change of law, as much as 69 TB^[http://www.gugik.gov.pl/aktualnosci/2021/03.09.2020-sierpniowe-statystyki-pobierania-uwolnionych-danych-przestrzennych] of data was downloaded, and by the end of October, this value grew to over 240 TB^[http://www.gugik.gov.pl/aktualnosci/03.11.2020-statystyka-pobierania-danych-w-pazdzierniku].
+In the first month after the change of law, 69 TB of data was downloaded^[http://www.gugik.gov.pl/aktualnosci/2021/03.09.2020-sierpniowe-statystyki-pobierania-uwolnionych-danych-przestrzennych], and by the end of October, this value grew to over 240 TB^[http://www.gugik.gov.pl/aktualnosci/03.11.2020-statystyka-pobierania-danych-w-pazdzierniku].
 
 # Statement of need
 
-While the Geoportal gives access to some of the *GUGiK* data resources, it has a number of practical disadvantages.
-Datasets can only be downloaded individually and manually, limiting the practical use of the available data for large areas or many points in time, and is problematic for the reproducible research process.
-Additionally, some *GUGiK* data is located on other related websites or in the form of dedicated services, which makes finding and downloading some datasets more difficult.
+While the Geoportal gives access to some of the *GUGiK* data resources, it has several practical disadvantages.
[…]
