### 2021-03-04 e1f4367
add 'err_print()' fun to 'utils.R'
