+++ R/utils.R
@@ -4,3 +4,9 @@ tryGet = function(code) {
            warning = function(c) "warning"
   )
 }
+
+err_print = function() {
+  msg = paste0("Conection error.", "\n", "If the problem persists, ",
+               "try a different download method from 'download.file()'.")
+  message(msg)
+}
