### 2021-03-04 f0d77d6
return better error message in '*_download' funs
