+++ R/topodb_download.R
@@ -70,7 +70,7 @@ topodb_download = function(county = NULL, TERYT = NULL, outdir = ".",
     status = tryGet(utils::download.file(prepared_URL, filename, mode = "wb", ...))
 
     if (any(status %in% c("error", "warning"))) {
-      return("connection error")
+      return(err_print())
     }
 
     if (unzip) {
