+++ R/borders_download.R
@@ -44,7 +44,7 @@ borders_download = function(type, outdir = ".", unzip = TRUE, ...) {
   status = tryGet(utils::download.file(URL, filename, mode = "wb", ...))
 
   if (any(status %in% c("error", "warning"))) {
-    return("connection error")
+    return(err_print())
   }
 
   if (unzip) {
