+++ R/emuia_download.R
@@ -61,7 +61,7 @@ emuia_download = function(commune = NULL, TERYT = NULL, outdir = ".",
     status = tryGet(utils::download.file(df_names[i, "URL"], filename, mode = "wb", ...))
 
     if (any(status %in% c("error", "warning"))) {
-      return("connection error")
+      return(err_print())
     }
 
     if (unzip) {
