### 2021-03-04 kadyb/rgugik#64 @kadyb
After a long discussion and many tests, we can propose a solution consisting of two steps:

1. Adding information about the possibility of using a different download method when a connection error occurs (*it's already done*). We use `utils::download.file()` function as a backend and it allows downloading using various methods / tools that can be selected by the user. During testing, we concluded that for most users the default method works fine, but there are cases where it fails (probably especially for foreign users). As you mentioned, choosing a different method (e.g. `wget`) may solve this problem. Thus, we recommend this in https://github.com/kadyb/rgugik/pull/68.

2. We plan to add `curl::curl_download()` as an additional independent download tool in the long term (after review). We did some tests and we think it will be a good solution, however we encountered a [bug on Windows](https://github.com/jeroen/curl/issues/238) which is holding us back at the moment. Nevertheless, this solution can also be tested by users in the development version of the package in this way: `remotes::install_github("kadyb/rgugik@usecurl")`.
