### 2021-03-04 d80a0fb
add 'err_print()' test
