+++ tests/testthat/test-utils.R
@@ -1,0 +1,1 @@
+expect_message(rgugik:::err_print(), "Conection error.")
