### 2021-03-04 #2948 @kadyb
Hello, @mikerspencer thanks for the review and praise! I would like to respond here to the downloading problem you described:

> (...) unfortunately I was unable to test functionality due to issue 64. Issue 64 is a connection error when downloading datasets, which consistently occurs at ~15 MB through a download. The issue doesn't affect smaller downloads, e.g. the 2.4 MB files in the automated tests. I've tested at two different times of day and am reasonably sure my internet connection is stable.

Both reviewers confirmed this problem. On our side it is hard to reproduce, but the error is certainly there. We think it rather affects foreign users (we suppose it could be due to lower server priority). Nevertheless, thanks to the very active and helpful attitude of the reviewers, we tested two solutions to prevent this problem:
1. All download functions in the package use the `utils::download.file()` backend, which supports different download methods. The default method failed, however reviewers confirmed that changing the method helps to resolve the issue. We have therefore corrected the error returned to clearly inform the user that they can change the method (https://github.com/kadyb/rgugik/pull/68).
2. In the long run (after review), we will add another independent download tool (`curl::curl_download()`) as a backend. However, we detected a problem on Windows (it has been [reported](https://github.com/jeroen/curl/issues/238) to the package developer) and it requires more testing to ensure smooth operation (user can test this using the development version of the package).

More information can be found [here](https://github.com/kadyb/rgugik/issues/64). We believe that the actions taken are sufficient.
