+++ man/models3D_download.Rd
@@ -40,6 +40,6 @@ Download 3D models of buildings for counties
 \examples{
 \dontrun{
 models3D_download(TERYT = c("2476", "2264")) # 3.6 MB
-models3D_download(county = "sejneĹ„ski", LOD = "LOD2") # 7.0 MB
+models3D_download(county = "sejneński", LOD = "LOD2") # 7.0 MB
 }
 }
