+++ man/topodb_download.Rd
@@ -29,7 +29,7 @@ Download Topographic Databases for counties
 }
 \examples{
 \dontrun{
-topodb_download(county = "ĹšwiÄ™tochĹ‚owice") # 2.4 MB
+topodb_download(county = "Świętochłowice") # 2.4 MB
 topodb_download(TERYT = c("2476", "2264")) # 4.8 MB
 }
 }
