### 2021-03-05 #2948 @mikerspencer
@kadyb good work on the solution. Which version should I test to complete my review? I presume the updated error returned in not yet on CRAN?
