### 2021-03-05 #2948 @kadyb
@mikerspencer, yes, you're right. Currently, all changes have been made to the development version on GitHub.

Please use the code below to install the development version (this will allow to test the first proposed solution):

```r
# install.packages("remotes")
remotes::install_github("kadyb/rgugik")
```

To test the second solution (downloading files via `curl` package) , please use this:

```r
remotes::install_github("kadyb/rgugik@usecurl")
```
