### 2021-03-10 kadyb/rgugik#64 @mikerspencer
Can you add the fix (e.g. `method = "wget"`) to the vignettes? It's going to be something that trips up a beginner.
