### 2021-03-10 kadyb/rgugik#69 @mikerspencer
vignette plotting error
The DEM vignette (https://github.com/kadyb/rgugik/blob/master/vignettes/DEM.Rmd) has a bug in the plotting at line 311. The plotting space retains settings from the DEM plots and needs `dev.off()`, or similar, to reset defaults.

Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/2948
