### 2021-03-10 kadyb/rgugik#69 @mikerspencer
Also affects Topo vignette (https://github.com/kadyb/rgugik/blob/master/vignettes/topodb.Rmd) at line 248. Likely a problem with the way the legend is plotted (i.e. subsequent plots appear in the legend space).
