### 2021-03-10 kadyb/rgugik#70 @mikerspencer
GDAL error: unit of measure not found
Ortho vignette (https://github.com/kadyb/rgugik/blob/master/vignettes/orthophotomap.Rmd), line 236/7.

    Warning message:
    In CPL_read_gdal(as.character(x), as.character(options), as.character(driver),  :
      GDAL Error 1: PROJ: proj_uom_get_info_from_database: unit of measure not found

My gdal version (3.2.1 on Ubuntu 20.04).

The following line (247) generates 38 gdal warnings:

    38: In CPL_read_gdal(as.character(x), as.character(options),  ... :
    GDAL Message 1: for band 3, nodata value has been clamped to 0,
      the original value being out of range.

The same out of range error appears from line 260 and 271, but the plots are created.

Submitted as part of JOSS review: https://github.com/openjournals/joss-reviews/issues/2948
