### 2021-03-10 kadyb/rgugik#70 @kadyb
Which version of `sf` and `stars` are you using?

How do you use these vignettes? From [website](https://kadyb.github.io/rgugik/articles/) and rewrite the code to R? Or do you run Rmarkdown from [vignettes catalog](https://github.com/kadyb/rgugik/tree/master/vignettes)?
