### 2021-03-10 c6e8748
mention 'dev.off()' to reset plots in vignettes
