### 2021-03-10 b031ccd
update R and regenerate vignette
