### 2021-03-10 kadyb/rgugik#69 @kadyb
Thanks! We generate vignettes from Rmarkdown files in which each figure is plotted in a separate chunk (in a new graphic device), so we didn't encounter a bug with incorrect plotting. Just as you suggested, `dev.off()` resolves the error for R scripts. I added such a note as a comment in https://github.com/kadyb/rgugik/pull/71.
