### 2021-03-10 kadyb/rgugik#70 @kadyb
I tried to reproduce this with `GEOS 3.8.0, GDAL 3.0.4, PROJ 6.3.1, sf 0.9.7, stars 0.5.2` on Windows 8.1 and Pop!_OS 20.04 LTS. This warning and GDAL message didn't show up for me, so it may be only related to the logging options.

It seems to me that there is nothing to worry about. The warning is related to the lack of specification of the coordinate system in the raster. We use `st_crs(img) = 2180` in next steps to fix it. The GDAL message also seems to be related to data, so I think it can be skipped.
