### 2021-03-10 kadyb/rgugik#64 @kadyb
I've thought about it, but I'm not convinced that setting this as the default for download is a universal solution. During testing, we noticed that different methods are optimal for different configurations. I think the best solution in this situation would be to leave the default method in `utils::download.file()` (`wget` isn't installed on Windows by default).

However, the new user needs to be aware that there are different download methods that can be used, so I think adding this information to the vignettes would be a great idea.
