+++ vignettes/orthophotomap.Rmd
@@ -67,6 +67,13 @@ Let's do that.
 geodb_download("wielkopolskie", outdir = "./data")
 ```
 
+If you run into any problem with the download, remember that you can pass another download method from `download.file()` as a function argument.
+
+
+```r
+geodb_download(req_df, outdir = "./data", method = "wget")
+```
+
 The downloaded database consists of many files in the GML (*Geography Markup Language*) format.
 A brief description of the structure of this database can be found [here](https://kadyb.github.io/rgugik/articles/articles/spatialdb_description.html).
 The table with the nature reserves is in the "PL.PZGIK.201.30__OT_TCRZ_A.xml" file.
@@ -136,6 +143,10 @@ We need to provide our Krajkowo polygon as the argument of this function.
 req_df = ortho_request(krajkowo)
 ```
 
+```
+## Error in ortho_request(krajkowo): could not find function "ortho_request"
+```
+
 We can display the resulting table using the code below.
 
 
@@ -145,17 +156,7 @@ req_df[1:10, 1:6]
 ```
 
 ```
-##             sheetID year resolution composition  sensor     CRS
-## 1  N-33-142-B-d-4-2 2004       0.50         B/W  Analog PL-1992
-## 2  N-33-142-B-d-4-4 2004       0.50         B/W  Analog PL-1992
-## 3  N-33-142-B-d-4-4 2010       0.25         RGB Digital PL-1992
-## 4  N-33-142-B-d-4-2 2010       0.25         RGB Digital PL-1992
-## 5  N-33-142-B-d-4-4 2010       0.25         CIR Digital PL-1992
-## 6  N-33-142-B-d-4-2 2010       0.25         CIR Digital PL-1992
-## 7  N-33-142-B-d-4-2 2016       0.25         CIR Digital PL-1992
-## 8  N-33-142-B-d-4-2 2016       0.25         RGB Digital PL-1992
-## 9  N-33-142-B-d-4-4 2016       0.25         CIR Digital PL-1992
-## 10 N-33-142-B-d-4-4 2016       0.25         RGB Digital PL-1992
+## Error in eval(expr, envir, enclos): object 'req_df' not found
 ```
 
 To complete our task, we need to obtain a near infrared data.
@@ -167,11 +168,19 @@ So in the next step, we select those rows for which the `composition` column has
 req_df = req_df[req_df$composition == "CIR", ]
 ```
 
+```
+## Error in eval(expr, envir, enclos): object 'req_df' not found
+```
+
 Then let's sort the table according to the year the photo was taken, with the most recent images at the top of the table.
 
 
 ```r
 req_df = req_df[order(-req_df$year), ]
+```
+
+```
+## Error in eval(expr, envir, enclos): object 'req_df' not found
 ```
 
 Let's display the table again and select the newest compositions.
@@ -182,18 +191,16 @@ req_df[, c(1:5, 9)]
 ```
 
 ```
-##             sheetID year resolution composition  sensor seriesID
-## 7  N-33-142-B-d-4-2 2016       0.25         CIR Digital    69837
-## 9  N-33-142-B-d-4-4 2016       0.25         CIR Digital    69837
-## 22 N-33-142-B-d-4-2 2013       0.25         CIR Digital    69903
-## 24 N-33-142-B-d-4-4 2013       0.25         CIR Digital    69903
-## 5  N-33-142-B-d-4-4 2010       0.25         CIR Digital    69763
-## 6  N-33-142-B-d-4-2 2010       0.25         CIR Digital    69763
+## Error in eval(expr, envir, enclos): object 'req_df' not found
[…]
