### 2021-03-10 kadyb/rgugik#64 @kadyb
I added such an explanation and example in: https://github.com/kadyb/rgugik/pull/72
