### 2021-03-11 kadyb/rgugik#70 @mikerspencer
sf version 0.9-7
stars version 0.4-3

I agree it's not a worry - everything worked subsequently and was plotted in the right place.
