### 2021-03-11 kadyb/rgugik#70 @kadyb
Right. I can also recommend upgrading `stars` to the new version because it is a really great package, but I don't know if it will affect the printing of warnings/messages from GDAL. Anyway, thanks for reporting!
