### 2021-03-14 #2948 @adamhsparks
Code all looks good. I've added comments in the PDF that I've attached here for my comments on the paper as requested.

[10.21105.joss.02948.pdf](https://github.com/openjournals/joss-reviews/files/6136046/10.21105.joss.02948.pdf)
