### 2021-03-15 #2948 @adamhsparks
The paper's grammar looks good now but the test coverage, both with the badge and with the paper is still an issue. I just pulled the latest and ran `covr::report()` and got this result of ~80% coverage. Not 87% or 89%?
