### 2021-03-15 #2948 @adamhsparks
Ah, thanks for the explanation.

I’m curious, why not use the `#nocov` tags in these files to exclude coverage if they aren’t/don’t need to be tested?

> On 15 Mar 2021, at 6:15 pm, Krzysztof Dyba ***@***.***> wrote:
>
>
> geonames_download() and borders_download() functions are just for data download and do not contain any complex logic. They are not automatically tested because the downloaded files are large (e.g., 375 MB) and server space is limited. We use this <https://github.com/kadyb/rgugik/blob/master/.github/workflows/test-coverage.yml> configuration file to calculate the test coverage value on codecov.io <https://codecov.io/gh/kadyb/rgugik>. Please run this code to get identical results locally:
>
> fun_excl = c("geonames_download", "borders_download")
> cov = covr::package_coverage(function_exclusions = fun_excl)
> cov
> #> rgugik Coverage: 87.55%
