### 2021-03-15 #2948 @kadyb
> I’m curious, why not use the `#nocov` tags in these files to exclude coverage if they aren’t/don’t need to be tested?

I could be wrong, but I think the preferred way is to use the `function_exclusions` argument. However, there are many ways to achieve the effect.

@adamhsparks, thanks for your time and helpful comments!
