### 2021-03-15 #2948 @adamhsparks
Thanks, in that case I may be doing it wrong! I’ve learned something new doing this review.

> On 15 Mar 2021, at 7:00 pm, Krzysztof Dyba ***@***.***> wrote:
>
>
> I’m curious, why not use the #nocov tags in these files to exclude coverage if they aren’t/don’t need to be tested?
>
> I could be wrong, but I think the preferred way is to use the function_exclusions argument. However, there are many ways to achieve the effect.
>
> @adamhsparks <https://github.com/adamhsparks>, thanks for your time and helpful comments!
>
