### 2021-03-15 #2948 @kadyb
1. https://github.com/kadyb/rgugik/releases/tag/v0.2.1 (version tag: `v0.2.1`)
2. https://zenodo.org/record/4606706
3. Done.
4. https://doi.org/10.5281/zenodo.4606705 (`10.5281/zenodo.4606706`)

If something is wrong, please let me know.
