### 2021-03-15 #2948 @kbarnhart
@kadyb I've now recommended that this submission be accepted and published. One of the JOSS editors in chief @openjournals/joss-eics will handle the submission from here. Congratulations.

Thanks to @adamhsparks and @mikerspencer for contributing thoughtful and thorough reviews.
