### 2021-03-16 #2948 @arfon
@kadyb - I've made a few formatting fixes and small language changes in https://github.com/kadyb/rgugik/pull/74 - please let me know what you think.

Also, the footnotes on the first page were overflowing the margin of the paper so I moved them to links. Hopefully this is OK? If not, an alternative would be to make these proper references in the BibTeX.
